Added two methods to the GasPump class.
The first method I added was selectPaymentType that allowed the user to
choose what way they wanted to pay and then I'd assign their choice to
a variable to use later. I also made a go method so that it is the only
one public to the user. This will be were I run the other methods once
I create them.